requirements1.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-common-sql==1.0.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-google==8.0.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.1.0
apispec==3.3.2
arabic-reshaper==2.1.3
argcomplete==2.0.0
asgiref==3.5.2
asn1crypto==1.5.1
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
Babel==2.10.3
black==22.3.0
blinker==1.5
boto==2.49.0
boto3==1.24.56
botocore==1.27.56
cachelib==0.9.0
cachetools==5.2.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
constantly==15.1.0
cron-descriptor==1.2.31
croniter==1.3.5
cryptography==36.0.1
cssselect2==0.6.0
daphne==3.0.2
db-dtypes==1.0.2
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1
django-autoslug==1.9.8
django-jazzmin==2.5.0
django-storages==1.13.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
email-validator==1.2.1
ez-setup==0.9
Flask==2.2.1
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.3
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
google-ads==18.0.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-aiplatform==1.16.1
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.3
google-cloud-automl==2.8.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.7.0
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==1.7.2
google-cloud-build==3.9.0
google-cloud-container==2.11.0
google-cloud-core==2.3.2
google-cloud-datacatalog==3.9.0
google-cloud-dataplex==1.1.0
google-cloud-dataproc==5.0.0
google-cloud-dataproc-metastore==1.6.0
google-cloud-dlp==1.0.2
google-cloud-kms==2.12.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.1
google-cloud-memcache==1.4.1
google-cloud-monitoring==2.10.1
google-cloud-orchestration-airflow==1.4.1
google-cloud-os-login==2.7.1
google-cloud-pubsub==2.13.4
google-cloud-redis==2.9.0
google-cloud-resource-manager==1.6.0
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==1.44.0
google-cloud-tasks==2.10.0
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.7.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-gcp==0.2.2
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
json-merge-patch==0.2
jsonschema==4.9.1
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
looker-sdk==22.10.0
lxml==4.9.0
Mako==1.2.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
num2words==0.5.10
numpy==1.23.0
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pandas-gbq==0.17.7
pathspec==0.9.0
paytmchecksum==1.7.0
pendulum==2.1.2
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
prison==0.2.1
proto-plus==1.19.6
protobuf==3.19.4
psutil==5.9.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.12.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyrsistent==0.18.1
python-bidi==0.4.2
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
qrcode==7.3.1
random2==1.0.1
razorpay==1.3.0
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
s3transfer==0.6.0
service-identity==21.1.0
setproctitle==1.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlalchemy-bigquery==1.4.4
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
svglib==1.3.0
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.2.1
wrapt==1.14.1
WTForms==2.3.3
xhtml2pdf==0.2.7
zope.interface==5.4.0
